requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.12.5
chardet==4.0.0
ci-info==0.2.0
cycler==0.10.0
decorator==4.4.2
Django @ file:///tmp/build/80754af9/django_1609784082548/work
djangorestframework==3.12.2
fonttools==4.18.2
idna==2.10
isodate==0.6.0
kiwisolver @ file:///D:/bld/kiwisolver_1610099953470/work
lxml==4.6.2
matplotlib @ file:///D:/bld/matplotlib-suite_1605180431619/work
networkx==2.5
numpy @ file:///D:/bld/numpy_1609881266195/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging==20.8
pandas==1.2.0
pdf2image==1.14.0
Pillow @ file:///D:/bld/pillow_1609628021885/work
psycopg2 @ file:///C:/ci/psycopg2_1608147654043/work
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
rdflib==5.0.0
requests==2.25.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tornado @ file:///D:/bld/tornado_1610094910414/work
urllib3==1.26.2
wincertstore==0.2
wordcloud @ file:///D:/bld/wordcloud_1605555768540/work
